pyproject.toml
[project]
name = "house-hunter"
version = "1.0.0"
description = "AI-powered real estate search agent using LangGraph and GPT-4o-mini"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "langsmith>=0.4.34",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "pytz>=2024.1",
]

[project.scripts]
house-hunter = "house_hunter.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["house_hunter"]
